Cleaned up ImGui implementation Assigned useful variables to ImGui and adjusted math for calculating rotation and scale. Implemented Circle vertex generation Implemented Circle Drawing (added new primatives and setup "type" switch case parameter.
TODO:
Need to fully comment and cleanup code.
Need to classify dynamic circle creation. Ran into an interesting C++ challenge with dynamically resizing arrays. I want to have a ImGui slider to change the circle radius and resolution but that means i'll need to destroy and reallocate the array whenever it's changed. need to handle this in the circle class.
Need to implement transforms on all globjects.
Need to check for poor cleanup causing crashes.... Not sure how to go about this.
Once cleanup is done.... will consider forking into mars game.